Move out code related to ESIGMA model waveform generation
This has now been moved to:
https://github.com/gwnrtools/esigmapy
